# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule:
    properties:
      class:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class'
      class_id:
        type: string
      day_of_week:
        description: 1 = Senin, 7 = Minggu (ISO 8601)
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      end_time:
        description: standard HH:MM:SS for time column
        type: string
      enrollment:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment'
      enrollment_id:
        type: string
      id:
        type: string
      location:
        type: string
      start_time:
        description: standard HH:MM:SS for time column
        type: string
      tutor_id:
        type: string
      valid_from:
        type: string
      valid_until:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateInitialSchedulesRequest:
    properties:
      class_id:
        type: string
      schedules:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ScheduleItemRequest'
        minItems: 1
        type: array
    required:
    - class_id
    - schedules
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateInitialSchedulesResponse:
    properties:
      schedules:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule'
        type: array
      sessions:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession'
        type: array
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse:
    properties:
      data: {}
      message:
        example: An error occurred
        type: string
      status:
        example: error
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.PermanentScheduleChangeRequest:
    properties:
      effective_date:
        type: string
      new_day_of_week:
        maximum: 7
        minimum: 1
        type: integer
      new_end_time:
        type: string
      new_start_time:
        type: string
      old_schedule_id:
        type: string
    required:
    - effective_date
    - new_day_of_week
    - new_end_time
    - new_start_time
    - old_schedule_id
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.PermanentScheduleChangeResponse:
    properties:
      new_schedule:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule'
      new_sessions:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession'
        type: array
      old_schedule:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule'
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.PermanentTutorChangeRequest:
    properties:
      effective_date:
        type: string
      new_tutor_id:
        type: string
      schedule_id:
        type: string
    required:
    - effective_date
    - new_tutor_id
    - schedule_id
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.PermanentTutorChangeResponse:
    properties:
      new_schedule:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule'
      old_schedule:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule'
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.RescheduleSessionRequest:
    properties:
      new_end_time:
        type: string
      new_location:
        type: string
      new_session_date:
        type: string
      new_start_time:
        type: string
      session_id:
        type: string
    required:
    - new_end_time
    - new_session_date
    - new_start_time
    - session_id
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.RescheduleSessionResponse:
    properties:
      new_session:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession'
      original_session:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession'
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ScheduleItemRequest:
    properties:
      day_of_week:
        description: 1=Monday, ..., 7=Sunday
        maximum: 7
        minimum: 1
        type: integer
      end_time:
        description: HH:MM:SS
        type: string
      enrollment_id:
        type: string
      location:
        type: string
      start_time:
        description: HH:MM:SS
        type: string
      tutor_id:
        type: string
      valid_from:
        type: string
    required:
    - day_of_week
    - end_time
    - start_time
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ScheduleListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule'
        type: array
      pagination:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination'
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.SessionAttendeesResponse:
    properties:
      attendees:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment'
        type: array
      class_id:
        type: string
      session_id:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.SubstituteTutorRequest:
    properties:
      session_id:
        type: string
      substitute_tutor_id:
        type: string
    required:
    - session_id
    - substitute_tutor_id
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.SubstituteTutorResponse:
    properties:
      session:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession'
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
  description: Academic Management Service for KelolaKelas Platform
  title: KelolaKelas Academic Service API
  version: "1.0"
paths:
...
  /api/v1/schedules:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ScheduleListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List tenant schedules
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      description: Create recurring schedule items and corresponding class sessions
        for a class
      parameters:
      - description: Tenant ID dalam format UUID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Create initial schedules request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateInitialSchedulesRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateInitialSchedulesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create initial schedules for a class
      tags:
      - Schedules
  /api/v1/schedules/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-delete a tenant-owned schedule and cancel its future scheduled
        sessions
      parameters:
      - description: Tenant ID dalam format UUID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Schedule ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete class schedule
      tags:
      - Schedules
  /api/v1/schedules/{id}/permanent:
    put:
      consumes:
      - application/json
      description: Apply permanent day/time schedule change and regenerate upcoming
        sessions
      parameters:
      - description: Tenant ID dalam format UUID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Schedule ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Permanent schedule change payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.PermanentScheduleChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.PermanentScheduleChangeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Permanently change schedule
      tags:
      - Schedules
  /api/v1/schedules/{id}/tutor-permanent:
    patch:
      consumes:
      - application/json
      description: Permanently reassign tutor for a schedule
      parameters:
      - description: Tenant ID dalam format UUID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Schedule ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Permanent tutor change payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.PermanentTutorChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.PermanentTutorChangeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Permanently change tutor
      tags:
      - Schedules
...
  /api/v1/sessions/{id}/attendees:
    get:
      consumes:
      - application/json
      description: Retrieve attendee list (enrolled students) for a specific session
      parameters:
      - description: Tenant ID dalam format UUID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Session ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.SessionAttendeesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get session attendees
      tags:
      - Sessions
  /api/v1/sessions/{id}/reschedule:
    post:
      consumes:
      - application/json
      description: One-off reschedule or make-up class for an existing session
      parameters:
      - description: Tenant ID dalam format UUID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Session ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Reschedule session payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.RescheduleSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.RescheduleSessionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reschedule a specific session
      tags:
      - Sessions
  /api/v1/sessions/{id}/substitute-tutor:
    patch:
      consumes:
      - application/json
      description: Assign a temporary substitute tutor for a single session
      parameters:
      - description: Tenant ID dalam format UUID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Session ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Substitute tutor payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.SubstituteTutorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.SubstituteTutorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign substitute tutor
      tags:
      - Sessions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"